Rename: 도메인명 변경 (Membership -> SquadMember)
