+++ src/main/java/com/eggmeonina/scrumble/common/exception/ErrorCode.java
@@ -32,7 +32,7 @@ public enum ErrorCode {
 	 */
 	MEMBER_OR_SQUAD_NOT_FOUND(HttpStatus.BAD_REQUEST, "회원 또는 스쿼드가 존재하지 않습니다."),
 	SQUAD_NOT_FOUND(HttpStatus.BAD_REQUEST, "스쿼드가 존재하지 않습니다."),
-	MEMBERSHIP_NOT_FOUND(HttpStatus.BAD_REQUEST, "스쿼드에 존재하는 회원이 아닙니다."),
+	SQUADMEMBER_NOT_FOUND(HttpStatus.BAD_REQUEST, "스쿼드에 존재하는 회원이 아닙니다."),
 	UNAUTHORIZED_ACTION(HttpStatus.FORBIDDEN, "권한이 없는 사용자입니다.");
 
 	private final HttpStatus status;
