--- src/main/java/com/eggmeonina/scrumble/common/exception/MembershipException.java
@@ -1,7 +1,0 @@
-package com.eggmeonina.scrumble.common.exception;
-
-public class MembershipException extends ExpectedException{
-	public MembershipException(ErrorCode errorCode) {
-		super(errorCode);
-	}
-}
