+++ src/main/java/com/eggmeonina/scrumble/common/exception/SquadMemberException.java
@@ -1,0 +1,7 @@
+package com.eggmeonina.scrumble.common.exception;
+
+public class SquadMemberException extends ExpectedException{
+	public SquadMemberException(ErrorCode errorCode) {
+		super(errorCode);
+	}
+}
