--- src/main/java/com/eggmeonina/scrumble/domain/membership/domain/MembershipRole.java
@@ -1,8 +1,0 @@
-package com.eggmeonina.scrumble.domain.membership.domain;
-
-import lombok.Getter;
-
-@Getter
-public enum MembershipRole {
-	LEADER, NORMAL
-}
