--- src/main/java/com/eggmeonina/scrumble/domain/membership/dto/SquadUpdateRequest.java
@@ -1,14 +1,0 @@
-package com.eggmeonina.scrumble.domain.membership.dto;
-
-import jakarta.validation.constraints.NotEmpty;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class SquadUpdateRequest {
-	@NotEmpty
-	private String squadName;
-}
