--- src/main/java/com/eggmeonina/scrumble/domain/membership/repository/MembershipRepositoryCustom.java
@@ -1,10 +1,0 @@
-package com.eggmeonina.scrumble.domain.membership.repository;
-
-import java.util.List;
-
-import com.eggmeonina.scrumble.domain.membership.dto.SquadResponse;
-
-public interface MembershipRepositoryCustom {
-
-	List<SquadResponse> findSquads(Long memberId);
-}
