+++ src/main/java/com/eggmeonina/scrumble/domain/squadmember/domain/SquadMemberRole.java
@@ -1,0 +1,8 @@
+package com.eggmeonina.scrumble.domain.squadmember.domain;
+
+import lombok.Getter;
+
+@Getter
+public enum SquadMemberRole {
+	LEADER, NORMAL
+}
