+++ src/main/java/com/eggmeonina/scrumble/domain/squadmember/dto/SquadUpdateRequest.java
@@ -1,0 +1,14 @@
+package com.eggmeonina.scrumble.domain.squadmember.dto;
+
+import jakarta.validation.constraints.NotEmpty;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class SquadUpdateRequest {
+	@NotEmpty
+	private String squadName;
+}
