+++ src/main/java/com/eggmeonina/scrumble/domain/squadmember/repository/SquadMemberRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.eggmeonina.scrumble.domain.squadmember.repository;
+
+import java.util.List;
+
+import com.eggmeonina.scrumble.domain.squadmember.dto.SquadResponse;
+
+public interface SquadMemberRepositoryCustom {
+
+	List<SquadResponse> findSquads(Long memberId);
+}
