+++ src/main/java/com/eggmeonina/scrumble/domain/squadmember/repository/SquadRepositoryCustom.java
@@ -1,0 +1,10 @@
+package com.eggmeonina.scrumble.domain.squadmember.repository;
+
+import java.util.Optional;
+
+import com.eggmeonina.scrumble.domain.squadmember.dto.SquadDetailResponse;
+
+public interface SquadRepositoryCustom {
+
+	Optional<SquadDetailResponse> findSquadAndMembers(Long squadId);
+}
